+++ tests/benchdnn_graph/mlp/mlp.cpp
@@ -205,7 +205,7 @@ int doit(const mlp_graph_spec_t *spec, res_t *res) {
     std::vector<std::vector<dnnl::graph::logical_tensor>> ins_vec, outs_vec;
     std::vector<dnnl::graph::compiled_partition> cp_vec;
 
-    for (int i = 0; i < partitions.size(); i++) {
+    for (size_t i = 0; i < partitions.size(); i++) {
         const auto par = partitions[i];
         if (!par.is_supported()) continue;
 
@@ -269,7 +269,7 @@ int doit(const mlp_graph_spec_t *spec, res_t *res) {
         tensors_out.emplace_back(tensor_out);
     }
     //execute partitions
-    for (int i = 0; i < cp_vec.size(); i++) {
+    for (size_t i = 0; i < cp_vec.size(); i++) {
         SAFE(execute_and_wait(cp_vec[i], tensors_in[i], tensors_out[i], res),
                 WARN);
     }
